+++ Assets/Materials/CRT.mat
@@ -55,10 +55,11 @@ Material:
     m_Floats:
     - _Blur_Offset: 0.003
     - _EnableExternalAlpha: 0
+    - _Scanline_Size: 150
     - _Scanlines_Speed: 0.002
     - _ZWrite: 0
     m_Colors:
-    - _Color: {r: 16, g: 16, b: 16, a: 1}
+    - _Color: {r: 8, g: 8, b: 8, a: 1}
     - _RendererColor: {r: 1, g: 1, b: 1, a: 1}
   m_BuildTextureStacks: []
   m_AllowLocking: 1
