--- Assets/Scripts/Game.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c82b5cb12d829ba41886c262c00f7954
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
