--- Assets/Scripts/Game/MazeGenerate.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: d02d8b001af69d843aa3c85027b4e7cb
