--- Assets/Scripts/Game/MinigameInteract.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 212a24c1fd6c8e44f9e9beaed0b4eaa4
