--- Assets/Scripts/Game/Points.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 5f71672fb161dfb4d8994d3c918ac3b7
