--- Assets/Scripts/Game/SoundManager.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: dcdb97620f6d1aa45947c98303c27fdb
