+++ Assets/Scripts/Maze.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c82b5cb12d829ba41886c262c00f7954
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
