+++ Assets/Scripts/Maze/HintPointer.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: d520b5a4cc98131468c84126489c3626
