+++ Assets/Scripts/Maze/MinigameInteract.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 212a24c1fd6c8e44f9e9beaed0b4eaa4
