+++ Assets/Scripts/Maze/MinigameManager.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: cd0b7992725db9940abfea37a46b0dec
