+++ Assets/Scripts/Maze/PrefabMinigame.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: a5c1464fd82e60242b6214a581b1089f
