+++ Assets/Scripts/Maze/ShaderSetting.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 8219641020381f84fa146d7157b46f1e
