+++ Assets/Settings/Build Profiles/New macOS Profile 1.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3546d8edbf4f461499194a25493c3304
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
