+++ Assets/Settings/Build Profiles/New macOS Profile 2.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dd54fa02027c2c94d933367ab61623d8
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
