+++ Assets/URP/no post processing.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 66e1ede64cf75b84ea134c093b01b9bf
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
